Add new Fibonacci sequence generator functions
Two methods are derived from the exact solution formula. The first one
uses the another form derived from the exact solution without computing
the square root of 5, but it causes overflow when computing F(91). The
second one modifies the way it computes (a + b) / 2, the overflow occurs
when computing F(92).